Add files via upload
This is a capstone project which is meant to analyse data on 911 calls. I have used python in jupyter notebook IDE to analyze the reasons, wrangle with the data and extract meaningful features and thus build some interesting data visualization graphs using Seaborn, plotly in python. I have used Numpy and Pandas to wrangle the data.